diag: Make DCI change not to access info from a dangling pointer
When handling a DCI packet response, the uid information is being obtained from data that is already freed. Make changes so that the information is not obtained using a dangling pointer. Also, adding the header information to the DCI packet response is very unreadable. Make changes to improve the readability of adding the header information to the DCI packet response.